refactor: package domain 오타 수정
